+++ src/providers/provider.interface.ts
@@ -1,0 +1,7 @@
+import { CompletionItemProvider, DocumentSelector } from "vscode";
+
+export interface PathIntellisenseProvider {
+  selector: DocumentSelector;
+  provider: CompletionItemProvider;
+  triggerCharacters?: string[];
+}