+++ CHANGELOG.md
@@ -1,15 +1,24 @@
 ## Changelog
 
+### v2.0.0
+
+- [Changed] Refactored provider structure
+- [Fixed] Multi root workspace support
+
 ### v1.4.2
+
 - [Hotfix] Fixed broken intellisense on windows
 
 ### v1.4.1
+
 - [Hotfix] Fresh rebuild
 
 ### v1.4.0
+
 - [Feature] Added setting to define how to handle absolute paths Fixes #45, #55
 
 ### v1.3.0
+
 - [Feature] Files from "files.exclude" are now ignored as well
 - [Feature] Basic ts config support
 - [Bug] Fixed #26
@@ -18,39 +27,50 @@
 - Perf. improvements
 
 ### v1.2.0
+
 - [Feature] Add Option to automatically add a slash after directory. Thank you @koelpinl #41
 - [Bug] Fixed autocompletion for names with dashes. Thank you @kasperekt #36
 
 ### v1.1.0
+
 - [Feature] Custom Mappings #19, #5
 - [Feature] Setting to enable display of hidden files #12
 - [Bug] Fixed #11
 
 ### v1.0.2
+
 - [Bug] Fixed #15
 - [Bug] Fixed #16
 
 ### v1.0.1
+
 - Fixed compatibility with VS Code 1.3
 
 ### v1.0.0
+
 - [Feature] Added .. on top of each suggestion, Removed trailing slash
 - [Bugs] Various bugs fixed
 
 ### v0.2.0
+
 - [Feature] Remove file extension within imports (configurable)
 
 ### v0.1.0
+
 - Added slash after folders, group by folder and files, file icon
 
-### v0.0.5 
+### v0.0.5
+
 - Fixed windows path issue
 
-### v0.0.3 
+### v0.0.3
+
 - Fixed unknown publisher issue
 
-### v0.0.2 
+### v0.0.2
+
 - Updated Readme.md
 
-### v0.0.1 
-- Initial 
+### v0.0.1
+
+- Initial